Turn on new mega menu everywhere
This change turns on the new mega menu (previously referred to as
"variation 1"), by removing the MEGA_MENU_VAR_1 flag and switching to
the new code everywhere.
