+++ cfgov/unprocessed/css/main.less
@@ -117,7 +117,6 @@
 @import (less) "organisms/info-unit-group.less";
 @import (less) "organisms/item-introduction.less";
 @import (less) "organisms/mega-menu.less";
-@import (less) "organisms/mega-menu-var-1.less";
 @import (less) "organisms/mortgage-performance-trends.less";
 @import (less) "organisms/post-preview.less";
 @import (less) "organisms/prefooter.less";
