+++ cfgov/unprocessed/js/organisms/Header.js
@@ -1,20 +1,8 @@
 // Required modules.
 import { checkDom, setInitFlag } from '../modules/util/atomic-helpers';
 import GlobalSearch from '../molecules/GlobalSearch';
-import MegaMenuOrig from '../organisms/MegaMenu';
+import MegaMenu from '../organisms/MegaMenu';
 
-/*
-  TODO: Remove when 2019/2020 Mega Menu modifications are finalized.
-  Global variables are set by a flag in the base template.
-*/
-let MegaMenu;
-import MegaMenuVar1 from '../organisms/MegaMenuVar1';
-
-if ( window.cfpb && window.cfpb.megaMenuVar1 ) {
-  MegaMenu = MegaMenuVar1;
-} else {
-  MegaMenu = MegaMenuOrig;
-}
 
 /**
  * Header
