+++ cfgov/unprocessed/js/organisms/MegaMenuDesktop.js
@@ -21,8 +21,6 @@ function MegaMenuDesktop( baseClass, menus ) {
 
   // Binded functions.
   const _handleTriggerClickBinded = _handleTriggerClick.bind( this );
-  const _handleTriggerOverBinded = _handleTriggerOver.bind( this );
-  const _handleTriggerOutBinded = _handleTriggerOut.bind( this );
   const _handleExpandBeginBinded = _handleExpandBegin.bind( this );
   const _handleCollapseEndBinded = _handleCollapseEnd.bind( this );
 
@@ -34,9 +32,6 @@ function MegaMenuDesktop( baseClass, menus ) {
 
   // Whether this instance's behaviors are suspended or not.
   let _suspended = true;
-
-  // Timeout for delayed events.
-  let _showDelay;
 
   /**
    * @returns {MegaMenuDesktop} An instance.
@@ -62,55 +57,14 @@ function MegaMenuDesktop( baseClass, menus ) {
     if ( _suspended ) { return; }
     const eventMap = {
       triggerClick: _handleTriggerClickBinded,
-      triggerOver:  _handleTriggerOverBinded,
-      triggerOut:   _handleTriggerOutBinded,
       expandBegin:  _handleExpandBeginBinded,
       collapseEnd:  _handleCollapseEndBinded
     };
 
     const currHandler = eventMap[event.type];
     if ( currHandler ) {
-      const delay = _calcEventDelay( event.type );
-      if ( delay > 0 ) {
-        _delayedEvent( currHandler, event, delay );
-      } else {
-        currHandler( event );
-      }
-    }
-  }
-
-  /**
-   * @param {string} type - The type of event to check.
-   * @returns {number} The amount to delay in milliseconds,
-   *   length is determined based on the event type and
-   *   whether the menu is active or not.
-   */
-  function _calcEventDelay( type ) {
-    let delay = 0;
-    if ( type === 'triggerClick' ) {
-      window.clearTimeout( _showDelay );
-    } else if ( type === 'triggerOver' ) {
-      if ( _activeMenu === null ) {
-        delay = 150;
-      } else {
-        delay = 50;
-      }
-    }
-
-    return delay;
-  }
-
-  /**
-   * Delay the broadcasting of an event by supplied delay.
-   * @param {Function} currHandler - Event handler.
-   * @param {Event} event - A FlyoutMenu event.
-   * @param {number} delay - Delay in milliseconds.
-   */
-  function _delayedEvent( currHandler, event, delay ) {
-    window.clearTimeout( _showDelay );
-    _showDelay = window.setTimeout( function() {
       currHandler( event );
-    }, delay );
+    }
   }
 
   /**
@@ -122,24 +76,6 @@ function MegaMenuDesktop( baseClass, menus ) {
     const menu = event.target;
     if ( menu.isAnimating() ) { return; }
     _updateMenuState( menu, event.type );
-  }
-
-  /**
-   * Event handler for when FlyoutMenu trigger is hovered over.
-   * @param {Event} event - A FlyoutMenu event.
-   */
-  function _handleTriggerOver( event ) {
-    this.dispatchEvent( 'triggerOver', { target: this } );
-    _updateMenuState( event.target, event.type );
-  }
-
-  /**
-   * Event handler for when FlyoutMenu trigger is hovered out.
-   */
-  function _handleTriggerOut() {
-    this.dispatchEvent( 'triggerOut', { target: this } );
-    // Clear any queued events to show the menu.
-    window.clearTimeout( _showDelay );
   }
 
   /**
@@ -169,12 +105,13 @@ function MegaMenuDesktop( baseClass, menus ) {
   }
 
   /**
-   * Event handler for when mouse is hovering.
-   * @param {MouseEvent} event - The hovering event.
+   * Event handler for when clicking on the body of the document.
+   * @param {MouseEvent} event - The click event.
    */
-  function _handleMove( event ) {
-    // If we've left the parent container of the current menu, close it.
-    if ( !_firstLevelDom.contains( event.target ) ) {
+  function _handleBodyClick( event ) {
+    // If we've clicked outside the parent of the current menu, close it.
+    if ( !_firstLevelDom.contains( event.target ) ||
+         event.target.parentNode.classList.contains( `${ baseClass }_content-1-list` ) ) {
       _updateMenuState( null, event.type );
     }
   }
@@ -187,32 +124,27 @@ function MegaMenuDesktop( baseClass, menus ) {
   function _updateMenuState( menu, type ) {
     if ( menu === null || _activeMenu === menu ) {
       // A menu is closed.
-      window.clearTimeout( _showDelay );
       _activeMenu.getTransition().animateOn();
       _activeMenu.collapse();
       _activeMenu = null;
-      _bodyDom.removeEventListener( 'mousemove', _handleMove );
-      _bodyDom.removeEventListener( 'mouseleave', _handleMove );
+
+      // Clean up listeners
+      _bodyDom.removeEventListener( 'click', _handleBodyClick );
     } else if ( _activeMenu === null ) {
       // A menu is opened.
       _activeMenu = menu;
       _activeMenu.getTransition().animateOn();
 
-      /* Mousemove needed in addition to mouseout of the trigger
-         in order to check if user has moved off the menu <ul> and not
-         just the <li> list items. */
-      _bodyDom.addEventListener( 'mousemove', _handleMove );
-      _bodyDom.addEventListener( 'mouseleave', _handleMove );
+      // Close the menu on click of the document body.
+      _bodyDom.addEventListener( 'click', _handleBodyClick );
+
       _activeMenu.expand();
     } else {
       // An open menu has switched to another menu.
       _activeMenu.getTransition().animateOff();
       _activeMenu.collapse();
       _activeMenu = menu;
-      if ( type === 'triggerOver' ) {
-        _activeMenu.getTransition().animateOff();
-        _activeMenu.expand();
-      }
+      _activeMenu.getTransition().animateOff();
     }
   }
 
@@ -237,9 +169,6 @@ function MegaMenuDesktop( baseClass, menus ) {
     if ( !_suspended ) {
       treeTraversal.bfs( _menus.getRoot(), _handleSuspendTraversal );
 
-      // Ensure body events were removed.
-      _bodyDom.removeEventListener( 'mousemove', _handleMove );
-      _bodyDom.removeEventListener( 'mouseleave', _handleMove );
       // Clear active menu.
       _activeMenu = null;
       _suspended = true;
